The whole workspace checks on macOS now
