+++ crates/hring-bingo-ktls/src/non_linux.rs
@@ -1,0 +1,3 @@
+pub(crate) fn main() -> color_eyre::Result<()> {
+    panic!("hring-bingo-ktls is only supported on Linux");
+}
